requirements.txt
pytumblr==0.1.*
discord.py==2.*
Pillow==10.*
numpy